moving over fix from sp-pnp-js #795, remove core.ts from odata by moving content to parsers.ts
