+++ packages/odata/src/odata.ts
@@ -1,5 +1,4 @@
 export * from "./caching";
-export * from "./core";
 export * from "./parsers";
 export * from "./pipeline";
 export * from "./queryable";
