+++ packages/odata/src/odatabatch.ts
@@ -1,4 +1,4 @@
-import { ODataParser } from "./core";
+import { ODataParser } from "./parsers";
 import { getGUID, FetchOptions } from "@pnp/common";
 
 export interface ODataBatchRequestInfo {
